#60 レビュー指摘を反映
