+++ pubspec.yaml
@@ -37,11 +37,9 @@ dependencies:
     sdk: flutter
   freezed_annotation: ^3.1.0
   go_router: ^17.1.0
-  hooks_riverpod: ^3.2.1
+  hooks_riverpod: ^3.0.3
   intl: ^0.20.2
-  path: ^1.9.1
   riverpod_annotation: ^4.0.2
-  sqlite3_flutter_libs: ^0.6.0+eol
   widgetbook: ^3.21.0
   widgetbook_annotation: ^3.10.0
   yaml: ^3.1.3
